testing compute type

+++ index.py
@@ -12,7 +12,7 @@ model_size = "large-v2"
 device = "cuda" if torch.cuda.is_available() else "cpu"
 print(device)
 # Run on GPU with FP16
-whisper = WhisperModel(model_size, device=device, compute_type="float32")
+whisper = WhisperModel(model_size, device=device, compute_type="float16")
 # or run on GPU with INT8
 # model = WhisperModel(model_size, device="cuda", compute_type="int8_float16")
 # or run on CPU with INT8
